Count the distinct number of field values that exist per the given field to group by

This is a wrapper function for count_distinct_values_per_group from class nbaR.MultimediaClient.

multimedia_count_distinct_values_per_group(group = NULL, field = NULL,
  returnType = "data.frame", ...)

Arguments

group
name of field in the multimedia object you want to group by, type:

field
name of field in the multimedia object, type:

returnType
Either 'list' or 'data.frame', defaults to 'data.frame'

...
additional parameters passed to count_distinct_values_per_group from class nbaR.MultimediaClient

Value

scalar

Details

See also endpoint /getDistinctValuesPerGroup

See also