Create template quarto project to get an NBA html page up and running

Copies a template.qmd, template.scss, and _quarto.yml into a Quarto project directory.

Usage

nba_init_quarto_proj(
  path = "PathToMyProject/MyNewProject",
  overwrite = FALSE,
  files = c("scientific.qmd", "basic.qmd", "custom.scss", "_quarto.yml", "_brand.yml",
    "annals-of-the-new-york-academy-of-sciences.csl", "references.bib", "nba-banner.png",
    "sanbi-logo-small.png", "terrestrial-ecosystems.csv"),
  rename = NULL
)

Arguments

path

The destination directory for the project followed by the project name

overwrite

Whether to overwrite existing files. Defaults to FALSE.

files

Character vector of files to copy. Defaults to all template files.

rename

Named character vector to rename files on copy, e.g., c("scientific.qmd" = "index.qmd")

Value

Invisibly returns the paths of copied files.