Metadata operations

Client class for nbaR.Metadata objects

MetadataClient

Format

An object of class R6ClassGenerator of length 24.

Fields

basePath

Stores url path of the request.

userAgent

Set the user agent of the request.

Methods

get_allowed_date_formats

Get allowed values for dates in queries; Queries with other formatted dates will result in a query error. Parameters:

... : additional parameters passed to httr::GET

Returns: character

get_controlled_list_phase_or_stage

Get allowed values for the field &#39;PhaseOrStage&#39; in a specimen document; See also endpoint /getControlledLists. Parameters:

... : additional parameters passed to httr::GET

Returns: character

get_controlled_list_sex

Get allowed values for the field &#39;Sex&#39; in a specimen document; See also endpoint /getControlledLists. Parameters:

... : additional parameters passed to httr::GET

Returns: character

get_controlled_list_specimen_type_status

Get allowed values for the field &#39;SpecimenTypeStatus&#39; in a specimen document; See also endpoint /getControlledLists. Parameters:

... : additional parameters passed to httr::GET

Returns: character

get_controlled_list_taxonomic_status

Get allowed values for the field &#39;TaxonomicStatus&#39; in specimen and taxon documents; See also endpoint /getControlledLists. Parameters:

... : additional parameters passed to httr::GET

Returns: character

get_controlled_lists

Get the names of fields for which a controlled vocabulary exists; Possible values for fields with controlled vocabularies can be queried with metadata/getControlledList/field. Parameters:

... : additional parameters passed to httr::GET

Returns: character

get_rest_services

List all available REST services and their parameters; Lists end point name, http method, response type, and URL. Parameters:

... : additional parameters passed to httr::GET

Returns: RestService

get_setting

Get the value of an NBA setting; All settings can be queried with /metadata/getSettings. Parameters:

name : name of setting

... : additional parameters passed to httr::GET

Returns: list

get_settings

List all publicly available configuration settings for the NBA; The value of a specific setting can be queried with metadata/getSetting/name. Parameters:

... : additional parameters passed to httr::GET

Returns: list

get_source_systems

Get the data sources from which the data was retrieved; Returns code and name of all source systems. Parameters:

... : additional parameters passed to httr::GET

Returns: SourceSystem