Table of protection level of threatened ecosystems.

Function to style a table with threat status on the horizontal rows and protection level on the vertical columns, with number of ecosystems or taxa that share those categories, with total and percentage çolumns added

Usage

nba_tbl_comb(DF, GROUP, THR, PRO, FILE = c("spatial", "csv"))

Arguments

DF

The data frame that contains the data

GROUP

The column that contains the name of the variable (the ecosystems or taxa names)

THR

The column name of the threat statuses

PRO

The column name of the protection levels

FILE

An indication if the input file is a map (spatial file with a geom column) or a csv/ normal dataframe.

Examples

thr_pro_tbl <- NBA_example_map_data |>
nba_tbl_comb(GROUP = P_EcosysType,
THR = threat_status,
PRO = protection_level,
FILE = "spatial")

thr_pro_tbl
#> <table class="table table-striped table-hover" style="font-size: 10px; font-family: arial; width: auto !important; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;color: black !important;background-color: white !important;border-top: 2px solid black; border-bottom: 2px solid black; text-align: left; font-weight: normal;font-weight: bold;text-align: center;">  </th>
#>    <th style="text-align:center;color: black !important;background-color: white !important;border-top: 2px solid black; border-bottom: 2px solid black; text-align: left; font-weight: normal;font-weight: bold;text-align: center;"> Not Protected </th>
#>    <th style="text-align:center;color: black !important;background-color: white !important;border-top: 2px solid black; border-bottom: 2px solid black; text-align: left; font-weight: normal;font-weight: bold;text-align: center;"> Poorly Protected </th>
#>    <th style="text-align:center;color: black !important;background-color: white !important;border-top: 2px solid black; border-bottom: 2px solid black; text-align: left; font-weight: normal;font-weight: bold;text-align: center;"> Moderately Protected </th>
#>    <th style="text-align:center;color: black !important;background-color: white !important;border-top: 2px solid black; border-bottom: 2px solid black; text-align: left; font-weight: normal;font-weight: bold;text-align: center;"> Well Protected </th>
#>    <th style="text-align:center;color: black !important;background-color: white !important;border-top: 2px solid black; border-bottom: 2px solid black; text-align: left; font-weight: normal;font-weight: bold;text-align: center;"> Total </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;background-color: white !important;background-color: white !important;font-weight: bold;"> Critically Endangered </td>
#>    <td style="text-align:center;background-color: white !important;background-color: rgba(169, 56, 0, 255) !important;"> 0 </td>
#>    <td style="text-align:center;background-color: white !important;background-color: rgba(168, 112, 1, 255) !important;"> 1 </td>
#>    <td style="text-align:center;background-color: white !important;"> 1 </td>
#>    <td style="text-align:center;background-color: white !important;"> 1 </td>
#>    <td style="text-align:center;background-color: white !important;background-color: white !important;"> 3 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;background-color: white !important;background-color: white !important;font-weight: bold;"> Endangered </td>
#>    <td style="text-align:center;background-color: white !important;background-color: rgba(230, 152, 0, 255) !important;"> 0 </td>
#>    <td style="text-align:center;background-color: white !important;background-color: rgba(255, 235, 176, 255) !important;"> 1 </td>
#>    <td style="text-align:center;background-color: white !important;"> 0 </td>
#>    <td style="text-align:center;background-color: white !important;"> 1 </td>
#>    <td style="text-align:center;background-color: white !important;background-color: white !important;"> 2 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;background-color: white !important;background-color: white !important;font-weight: bold;"> Vulnerable </td>
#>    <td style="text-align:center;background-color: white !important;background-color: white !important;"> 1 </td>
#>    <td style="text-align:center;background-color: white !important;background-color: white !important;"> 0 </td>
#>    <td style="text-align:center;background-color: white !important;"> 1 </td>
#>    <td style="text-align:center;background-color: white !important;"> 1 </td>
#>    <td style="text-align:center;background-color: white !important;background-color: white !important;"> 3 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;background-color: white !important;background-color: white !important;font-weight: bold;"> Near Threatened </td>
#>    <td style="text-align:center;background-color: white !important;background-color: white !important;"> 0 </td>
#>    <td style="text-align:center;background-color: white !important;background-color: white !important;"> 0 </td>
#>    <td style="text-align:center;background-color: white !important;"> 1 </td>
#>    <td style="text-align:center;background-color: white !important;"> 1 </td>
#>    <td style="text-align:center;background-color: white !important;background-color: white !important;"> 2 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;background-color: white !important;background-color: white !important;font-weight: bold;"> Least Concern </td>
#>    <td style="text-align:center;background-color: white !important;background-color: white !important;"> 0 </td>
#>    <td style="text-align:center;background-color: white !important;background-color: white !important;"> 1 </td>
#>    <td style="text-align:center;background-color: white !important;"> 1 </td>
#>    <td style="text-align:center;background-color: white !important;"> 1 </td>
#>    <td style="text-align:center;background-color: white !important;background-color: white !important;"> 3 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;background-color: white !important;color: black !important;background-color: white !important;border-top: 2px solid black; border-bottom: 2px solid black;background-color: white !important;background-color: white !important;font-weight: bold;"> Total (n) </td>
#>    <td style="text-align:center;background-color: white !important;color: black !important;background-color: white !important;border-top: 2px solid black; border-bottom: 2px solid black;background-color: white !important;background-color: white !important;"> 1 </td>
#>    <td style="text-align:center;background-color: white !important;color: black !important;background-color: white !important;border-top: 2px solid black; border-bottom: 2px solid black;background-color: white !important;background-color: white !important;"> 3 </td>
#>    <td style="text-align:center;background-color: white !important;color: black !important;background-color: white !important;border-top: 2px solid black; border-bottom: 2px solid black;background-color: white !important;"> 4 </td>
#>    <td style="text-align:center;background-color: white !important;color: black !important;background-color: white !important;border-top: 2px solid black; border-bottom: 2px solid black;background-color: white !important;"> 5 </td>
#>    <td style="text-align:center;background-color: white !important;color: black !important;background-color: white !important;border-top: 2px solid black; border-bottom: 2px solid black;background-color: white !important;background-color: white !important;"> 13 </td>
#>   </tr>
#> </tbody>
#> </table>